restructured private inner clasess to separate packages.
