+++ community/cypher/cypher/src/main/scala/org/neo4j/cypher/internal/compatibility/v2_3/Compatibility.scala
@@ -33,8 +33,9 @@ import org.neo4j.cypher.internal.spi.v2_3.{TransactionBoundGraphStatistics, Tran
 import org.neo4j.cypher.internal.spi.v3_2.TransactionalContextWrapper
 import org.neo4j.graphdb.{Node, Relationship}
 import org.neo4j.kernel.GraphDatabaseQueryService
-import org.neo4j.kernel.api.ExecutingQuery.PlannerInfo
-import org.neo4j.kernel.api.{IndexUsage, KernelAPI}
+import org.neo4j.kernel.api.query.PlannerInfo
+import org.neo4j.kernel.api.KernelAPI
+import org.neo4j.kernel.api.index.IndexUsage
 import org.neo4j.kernel.impl.core.NodeManager
 import org.neo4j.kernel.impl.query.QueryExecutionMonitor
 import org.neo4j.kernel.monitoring.{Monitors => KernelMonitors}
