+++ community/kernel/src/main/java/org/neo4j/kernel/api/query/WaitingOnLock.java
@@ -1,0 +1,67 @@
+/*
+ * Copyright (c) 2002-2017 "Neo Technology,"
+ * Network Engine for Objects in Lund AB [http://neotechnology.com]
+ *
+ * This file is part of Neo4j.
+ *
+ * Neo4j is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+package org.neo4j.kernel.api.query;
+
+import java.util.HashMap;
+import java.util.Map;
+import java.util.concurrent.TimeUnit;
+
+import org.neo4j.storageengine.api.lock.ResourceType;
+import org.neo4j.time.SystemNanoClock;
+
+public class WaitingOnLock extends ExecutingQueryStatus
+{
+    private final String mode;
+    private final ResourceType resourceType;
+    private final long[] resourceIds;
+    private final long startTimeNanos;
+
+    public WaitingOnLock( String mode, ResourceType resourceType, long[] resourceIds, long startTimeNanos )
+    {
+        this.mode = mode;
+        this.resourceType = resourceType;
+        this.resourceIds = resourceIds;
+        this.startTimeNanos = startTimeNanos;
+    }
+
+    @Override
+    public long waitTimeNanos( SystemNanoClock clock )
+    {
+        return clock.nanos() - startTimeNanos;
+    }
+
+    @Override
+    public Map<String,Object> toMap( SystemNanoClock clock )
+    {
+        Map<String,Object> map = new HashMap<>();
+        map.put( "state", "WAITING" );
+        map.put( "lockMode", mode );
+        map.put( "waitTimeMillis", TimeUnit.NANOSECONDS.toMillis( waitTimeNanos( clock ) ) );
+        map.put( "resourceType", resourceType.toString() );
+        map.put( "resourceIds", resourceIds );
+        return map;
+    }
+
+    @Override
+    public boolean isPlanning()
+    {
+        return false;
+    }
+}
